Update DashGridLayout.react.js
Added new props from the original react component margin and allowOverlap. Added a new prop draggableChildStyle, with which you can modify the style of draggablewrappers when they are in edit mode. Default value is the same as it used to be in the corresponding css, which was removed from DashGridLayout.css